Implemented QWK(e) message support.
It's an exchange format and only basic reading is really supported. It's very close to pcboard so it can share some infrastructure.
It's still in use today and I can get QWK packages for testing & improving my other message bases.

BIG API change: Switched to the bstr crate for string representation. The issue when working with old data (like the PCBOARD qwk data in tests) is that there are
encoding issues - CP437 vs. UTF8. Encountering CP437 data shouldn't crash the message library. How to display the data at the end is up to the user of the library.

bstr is performant I could optimize loading with that as well.